requirements.txt
pandas
tabulate
streamlit>=1.24.0
pyswisseph>=2.10